Implemnetation of user management screen. Start of group management functions

+++ admin/controllers/groups/index.js
@@ -4,7 +4,9 @@ angular.module('santedb').controller('GroupIndexController', ["$scope", "$rootSc
      * @summary Delete the specified user
      */
     $scope.delete = function (id, index) {
-        if (confirm(SanteDB.locale.getString("ui.admin.groups.confirmDelete"))) {
+        var data = $("#SecurityRoleTable table").DataTable().row(index).data();
+
+        if (!data.obsoletionTime && confirm(SanteDB.locale.getString("ui.admin.groups.confirmDelete"))) {
             $("#action_grp_" + index + " a").addClass("disabled");
             $("#action_grp_" + index + " a i.fa-trash").removeClass("fa-trash").addClass("fa-circle-notch fa-spin");
             SanteDB.resources.securityRole.deleteAsync(id)
@@ -12,6 +14,38 @@ angular.module('santedb').controller('GroupIndexController', ["$scope", "$rootSc
                     $("#SecurityRoleTable table").DataTable().draw();
                 })
                 .catch($rootScope.errorHandler);
+        }
+        else if(data.obsoletionTime && confirm(SanteDB.locale.getString("ui.admin.groups.confirmUnDelete"))) {
+            $("#action_grp_" + index + " a").addClass("disabled");
+            $("#action_grp_" + index + " a i.fa-trash-restore").removeClass("fa-trash-restore").addClass("fa-circle-notch fa-spin");
+            
+            // Patch the user
+            var patch = new Patch({
+                change: [
+                    new PatchOperation({
+                        op: PatchOperationType.Remove,
+                        path: 'obsoletionTime',
+                        value: null
+                    }),
+                    new PatchOperation({
+                        op: PatchOperationType.Remove,
+                        path: 'obsoletedBy',
+                        value: null
+                    })
+                ]
+            });
+
+            SanteDB.resources.securityRole.patchAsync(id, null, patch)
+                .then(function (e) {
+                    $("#SecurityRoleTable").attr("newQuery", true);
+                    $("#SecurityRoleTable table").DataTable().draw();
+                })
+                .catch(function(e) {
+                    $("#action_grp_" + index + " a").removeClass("disabled");
+                    $("#action_grp_" + index + " a i.fa-circle-notch").removeClass("fa-circle-notch fa-spin").addClass("fa-trash-restore");
+                    $rootScope.errorHandler(e);
+                });
+
         }
     }
